requirements.txt
django
psycopg2
djangorestframework
pandas
mongoengine
pytest
pytest-django
mongomock